sgpp coding days
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2702 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/base/grid/storage/hashmap/HashGridStorage.hpp
@@ -586,7 +586,7 @@ public:
 	unsigned int store(index_pointer index)
 	{
 		list.push_back(index);
-		return (map[index] = static_cast<unsigned int>(this->seq() - 1));
+		return static_cast<unsigned int>(map[index] = static_cast<unsigned int>(this->seq() - 1));
 	}
 
 	/**
